i#23187: Fix crash of the document.
The mbLayoutInProgress bool was effectively unused - only set and reset, but
the only place that was checking for that was in lcl_RecalcRow(), again, only
to set and reset it.

Worse - with the document from i#23187, the mbLayoutInProgress was set / reset
on a page already disposed in SwFrm::InsertPage() which was causing the crash
here.

So let's get rid of all this mbLayoutInProgress, its getter and setter, and
NotifyLayoutOfPageInProgress class.

Valgrind, thank you that you exist, would have never found it without you! :-)

Change-Id: I9a9b604f8a21bb50e21af1b20b8db76d2c0504bd