sw: fix frame size of textboxes in DOCX/RTF export
In case the frame is AutoSize and the nominal height is less than the
real height, we used to end up with textboxes having height small enough
that some of the content was unreadable.

Instead, do what the WW8 export does and relay on sw::Frame to provide
the layout size.

Change-Id: I2a6cf4373c8565eef780273745a6ef27ddc65753