adapt code to data form .ui conversion
this dialog has some bizarre behaviour which is unchanged
with the .ui conversion

a) the scrollbar does not scroll the window beside it
up and down through the rows (which map to the spreadsheet columns),
but instead scrolls through the records (which map to the spreadsheet rows)
b) the dialog grows to include a row for each spreadsheet column, i.e.
large number of spreadsheet columns -> unusable dialog with too many rows
to fit on the screen.

what's indicated here is to clip the dialog to some max size and make the
scrollbar then scroll through the rows of the widget, and add another
scrollbar if necessary to scroll through the records.

Change-Id: I9b56992fc57468eb058d2a2914c08074f958a692