Revert "Calc Sort: Introducing new method AddKeyItem to avoid code duplication"
This reverts commit 3f41531f291f8d8f22a86430d11ae6d509a8ec09.

In tpsort.cxx this is used in another way than the other two cases:

++nSortKeyCount;
maSortKeyCtrl.AddSortKey( nSortKeyCount );
maSortKeyItems[nSortKeyIndex].m_pLbSort->SetSelectHdl( LINK( this, ScTabPageSortFields, SelectHdl ) );

So both are using the same index. In contrast the other two cases use (i+1)
and (i). So this can't be handled with one method or needs some more work...