Turn on _GLIBCXX_FULLY_DYNAMIC_STRING, too, in a dbgutil build
Apparently Apple's build of GNU libstdc++ has been compiled with that,
which means that also code using it should be. At least when
_GLIBCXX_DEBUG is also defined.

See http://lists.apple.com/archives/cocoa-dev/2009/Sep/msg01199.html and
http://stackoverflow.com/questions/2234557/c-using-getline-prints-pointer-being-freed-was-not-allocated-in-xcode

Otherwise the mork unit test crashes, because it happens to use
std::string in a way that triggers the problem. But quite likely there
would be more problems elsewhere.

Change-Id: Ie0c4fdfa61764718f333ba81015ef764806178fb