bring some sense into the error bar mappings for styles export
There were several properties that were duplicated for the X error bars
but where effectively not used by the code as these properties did not
work.

We now specify only one used entry from the chart2 API and remove the
entries from the old chart API. As soon as the old chart API is unused
in the import/export filter we can deprecate it.

Change-Id: Ib0487b41ea044136714037c27418ce049a5b92b7