Exif reading/writing and Jpeg lossless transformations support.
Support for reading/writing of Exif image metadata. Currently only
orientation is implemented, but support for other tags can be added.

Jpeg lossless transformations - currently only lossless rotation is
supported, but others can be added when needed.

Additionally GraphicNativeTransform and GraphicNativeMetadata has
been added. The responsibillity of GraphicNativeTransform is to
provide graphic transformations (like rotation) on native data and
the purpose is to be as lossless as possible in transformations.
GraphicNativeMetadata is a class for reading metadata which is
contained in a native data graphic stream. For now both support
only Jpeg.

Change-Id: I3e67cd3e7f5386746bcd1f0bfd2b90f5fe834b92