gbuild: work around problems with aux-targets and DLL renames (MSVC)
When renaming a DLL but not the import lib it can happen that
the "touch" rule creates the DLL file which is not helpful.
Since there is no easy way to fix this properly, and DLLs are not
renamed often, put in a lame hack to detect the situtation and delete
the import lib, which will get it re-linked on the next make invocation.

Change-Id: I3909cf7d377433d263e517719cca07005bade512