Generate new format type rdb files
With the main components (theTypeDescriptionManager, codemakers) supporting the
new unoidl format, it is time to switch the generated type rdb files to that new
format.

For now, this is done as a two-stage process, though:  Every
$(WORKDIR)/UnoApiTarget/*.rdb file in the new format is accompanied by an
*.rdb.oldformat file; both are generated in the same [UNO] gbuild step.  This is
for several reasons:

* idlc still generates old format output.

* The regcompare step (comparing against a reference rdb to catch
  incompatible changes) is still based on the old format.

(The reg2unoidl tool needs any dependend rdbs as additional inputs, that's why
the use_api information needs to be tracked as UNOAPI_DEPRDBS for now.  That can
be removed again when reg2unoidl is no longer used.)

Change-Id: Id625c88f9ecdbaba9e2af7410417e00b4ba36acc