fix scrolling performance of listboxes with hundreds of entries
make it approximately infinitely faster

a) add a method AdjustEntryHeightAndRecalc that just does AdjustEntryHeight and
RecalcViewData and replace all the places that do the two of those together
with that utility

b) in our SetFont now only call AdjustEntryHeightAndRecalc if the new font is
different to the old font, ignoring color which doesn't have a bearing on row
size.

c) where we originally called SetFont(GetFont()), which would have triggered
AdjustEntryHeightAndRecalc, now call the new AdjustEntryHeightAndRecalc
explicitly.

The performance problem apparently a regression from
1aefd8e86b81617e9204e6dc8a0a53d68740fd90 where we seem to now end up drawing on
ourself and calling SetFont on ourself rather than on a pImpl without an
overridden SetFont. So when we redraw on scrolling we triggered a cascade of
calculations and force text layout of every row.

Change-Id: I3c284c360f06ac383e8c38045fc4c2946f505b35