install scripts from shell using filelists
The FILELIST install method is really tailored to large sets of closely
related files. It is not such a great idea to apply it just to move some
unrelated files, delivered from a single module, out of $(OUTDIR), like
here, because it requires splitting one Package to several to allow the
files to be placed to different installation modules in scp2. The extra
makefile increases the overhead needed to place a file into an
installation set. We really need a better way to handle this...

Change-Id: I2f271562d8773152e69d284b4fe8ae356dea0945