fix the internal structure of MacOS X dmg
Because of commit 952373ed1bae92f9a0d0e3b906fe3b2a2242f464 "scp: remove
unused macros", which removed the top-level Applications symlink, the
dmg contained an extra dir, e.g.,
LibreOfficeDev_4.1.0.0.alpha0_MacOS_x86/LibreOfficeDev.app instead of
just LibreOfficeDev.app . This symlink used by the installer to
determine the dir to use as top-level for the installation
(solenv/bin/modules/installer/simplepackage.pm:274).

Also revert "Eek, it is hard to get the right path to the .app that is
being built", which was only needed because of the other problem.

This reverts commit 88c394a4e1b7d4739c60e833d8d88c00e1ab1b70.

Change-Id: Ib5792c7710decd021a62cd33fd841ebe26613fbc