Imrove scrolling in the Android "Desktop" app
I got scrolling to respond to the JNI scroll method call.  (This is not yet
the gesture I want to implement (scroll by pixels), but at least it’s a
start.)

At first I tried to do some hacking in the C++ part, then found the problem
was in the Java parameters, which set the mouse whell delta too big.  When
passed the value 1 or -1 to the Y value of AppSupport.scroll(int x, int y),
the scrolling is reasonable.

Change-Id: Id8fa0ada1a035760b7b05bf21325d0e51ef28fdc