Make harfbuzz Windows XP compatible
In a case we use harfbuzz on Windows once, it is using in the
uniscribe backend some OpenType shaping that is available only
in Vista and later. So call this functions over function pointers
if available and if not, fallback to the itemizing/shaping/placing
that is available in Windows XP uniscribe.

Change-Id: I5b1a22d812d57194d1147effa87033cdf6af78d3