Revert 1a3d6a0d9be4b5c87e90e351c129e903dd15caca
Which in itself was effectively a revert of
d3d56b104375cbb129c4694033a1e55da18cb6ef

Keeping the old broken line height calculation code is just masking of
the real problem; there are some code elsewhere that have fragile
workarounds to the real bug here (the removed code here shows a good
example of such workarounds). On Mac we use the correct metrics as well,
so we need to find the quirks and fix them, instead of pretending they
do not exist.

This fixes fdo#55469, among others.

Change-Id: I36f13b28eaba022b7c388feae7e0bfd0ed1c3e89