fixed transex3 out of bounds access to vMergeEntryVector
transex3 crashed if a new language was added to the StringArray
PairedList in svtools/source/misc/langtab.src and that resource was
attempted to be built.

For the new entry doesn't exist a MergeEntrys entry in vMergeEntryVector
yet, but ExportList::GetSourceLanguageListEntryCount() already includes
it so nMaxList is greater than size of vMergeEntryVector. Check nLIndex
first.

Change-Id: I3b706bea7aea7cd7ecd8245d99587b944510952d