Fix FreeType version check
FreeType has 3 different kinds of versions
* release, like 2.4.10
* libtool, like 13.0.7 (this what pkg-config returns)
* soname

FreeType's docs/VERSION.DLL provides a table mapping between the three

We were checking for >= 2.0 which is always true of course, so I
replaced it with 9.4.3 i.e. 2.1.5; the first FreeType release to provide
a freetype2.pc file (I could have omitted the version check at all,
since it is practically the same), but this should be replaced with a
real minimum required version.

Change-Id: I95f6c398a5b19d4bb728a6dddc957a9f95f27efa