Collect all empty broadcasters and remove them in one go later.
This is to avoid repeated calls to set_empty() on the broadcaster array
which causes array reallocation on each call. Instead, we can store the
segments of broadcasters to be deleted later, and call set_empty() on
those segments to reduce the number of calls to set_empty().

The meat of this is not implemented yet, as EndListeningContext is still
empty.

Change-Id: I914bc2881acee3902c4fec0f6c21aaf4d2489df8